Fix WPML tests
