+++ tests/test-timber-menu.php
@@ -356,7 +356,7 @@ class TestTimberMenu extends Timber_UnitTestCase {
 		$this->assertTrue( $menu_items[3]->current );
 		$this->assertContains( 'current-menu-item', $menu_items[3]->classes );
 	}
-  
+
 	function testMenuOptions () {
 		$menu_arr = self::_createTestMenu();
 
@@ -429,20 +429,6 @@ class TestTimberMenu extends Timber_UnitTestCase {
 		$this->assertEquals( 'http://upstatement.com', $item->url );
 		$this->assertTrue( $item->is_external() );
 
-	}
-
-	function testWPMLMenu() {
-		self::setPermalinkStructure();
-		self::_createTestMenu();
-		$menu = Timber::get_menu();
-		$nav_menu = wp_nav_menu( array( 'echo' => false ) );
-		$this->assertGreaterThanOrEqual( 3, count( $menu->get_items() ) );
-		$items = $menu->get_items();
-		$item = $items[0];
-		$this->assertEquals( 'home', $item->slug() );
-		$this->assertFalse( $item->is_external() );
-		$this->assertEquals( 'http://example.org/home/', $item->link() );
-		$this->assertEquals( '/home/', $item->path() );
 	}
 
 	function testMenuOptionsInNavMenuCssClassFilter() {
